Refactor: Modify folder name
